general: fix broken lightmaps due to multiple func_groups
So turns out my memory was incorrect and I never bothered to test this before committing to this strategy. Using multiple adjacent 'func_group' entities is *not* okay with lightmapping even if the shaders on the surfaces have 'q3map_lightmapmergable'.

This fixes lightmapping on all terrains by combining all the individual func_groups into a single func_group entity. This is accompanied by some special "reverse" shaders, to account for the differences in blending between PCX alphamaps vs 'q3map_alphaMod'. Because the former doesn't care about the blend direction, only that there's a shader that can do A -> B blend, there are some impossible to achieve blends in the maps when using 'q3map_alphaMod' instead. The solution is to create some reversed shaders (e.g. 1to2 -> 2to1) to fix the blend issues on these spots.